Move SpanFinishingCallback to top-level class
