+++ instrumentation/rocketmq/rocketmq-client/rocketmq-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rocketmqclient/v5_0/RocketMqConsumerInstrumentation.java
@@ -11,19 +11,12 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import apache.rocketmq.v2.ReceiveMessageRequest;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.api.internal.InstrumenterUtil;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import java.util.List;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.rocketmq.client.apis.message.MessageView;
 import org.apache.rocketmq.client.java.impl.consumer.ReceiveMessageResult;
-import org.apache.rocketmq.client.java.message.MessageViewImpl;
-import org.apache.rocketmq.shaded.com.google.common.util.concurrent.FutureCallback;
 import org.apache.rocketmq.shaded.com.google.common.util.concurrent.Futures;
 import org.apache.rocketmq.shaded.com.google.common.util.concurrent.ListenableFuture;
 import org.apache.rocketmq.shaded.com.google.common.util.concurrent.MoreExecutors;
@@ -59,66 +52,9 @@ final class RocketMqConsumerInstrumentation implements TypeInstrumentation {
         @Advice.Argument(0) ReceiveMessageRequest request,
         @Advice.Enter Timer timer,
         @Advice.Return ListenableFuture<ReceiveMessageResult> future) {
-      SpanFinishingCallback spanFinishingCallback = new SpanFinishingCallback(request, timer);
+      ReceiveSpanFinishingCallback spanFinishingCallback =
+          new ReceiveSpanFinishingCallback(request, timer);
       Futures.addCallback(future, spanFinishingCallback, MoreExecutors.directExecutor());
-    }
-  }
-
-  public static class SpanFinishingCallback implements FutureCallback<ReceiveMessageResult> {
-
-    private final ReceiveMessageRequest request;
-    private final Timer timer;
-
-    public SpanFinishingCallback(ReceiveMessageRequest request, Timer timer) {
-      this.request = request;
-      this.timer = timer;
-    }
-
-    @Override
-    public void onSuccess(ReceiveMessageResult receiveMessageResult) {
-      List<MessageViewImpl> messageViews = receiveMessageResult.getMessageViewImpls();
-      // Don't create spans when no messages were received.
-      if (messageViews.isEmpty()) {
-        return;
-      }
-      String consumerGroup = request.getGroup().getName();
-      for (MessageViewImpl messageView : messageViews) {
-        VirtualFieldStore.setConsumerGroupByMessage(messageView, consumerGroup);
-      }
-      Instrumenter<ReceiveMessageRequest, List<MessageView>> receiveInstrumenter =
-          RocketMqSingletons.consumerReceiveInstrumenter();
-      Context parentContext = Context.current();
-      if (receiveInstrumenter.shouldStart(parentContext, request)) {
-        Context context =
-            InstrumenterUtil.startAndEnd(
-                receiveInstrumenter,
-                parentContext,
-                request,
-                null,
-                null,
-                timer.startTime(),
-                timer.now());
-        for (MessageViewImpl messageView : messageViews) {
-          VirtualFieldStore.setContextByMessage(messageView, context);
-        }
-      }
-    }
-
-    @Override
-    public void onFailure(Throwable throwable) {
-      Instrumenter<ReceiveMessageRequest, List<MessageView>> receiveInstrumenter =
-          RocketMqSingletons.consumerReceiveInstrumenter();
-      Context parentContext = Context.current();
-      if (receiveInstrumenter.shouldStart(parentContext, request)) {
-        InstrumenterUtil.startAndEnd(
-            receiveInstrumenter,
-            parentContext,
-            request,
-            null,
-            throwable,
-            timer.startTime(),
-            timer.now());
-      }
     }
   }
 }
